Enemy movement, and damage implemented
