+++ Assets/Settings/Build Profiles.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 34c0e63ccd182c64dad75bf0a7e5a0c0
+guid: 51c66e5f1b1d445f2861309379fbf64d
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
