+++ Assets/Settings/Build Profiles/New Web Profile.asset.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 17b31a17764dd57479698f9787d395b6
-DefaultImporter:
+guid: 06e84cbbae7eb401091281a94f5b631c
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
