+++ ProjectSettings/TagManager.asset
@@ -6,6 +6,7 @@ TagManager:
   tags:
   - Cam
   - Enemy
+  - Bullet
   layers:
   - Default
   - TransparentFX
